==> 07-remove_gene_signature.Rmd <==
---
title: "Remove Gene Signature"
author: "Ramon Massoni-Badosa"
date: "2019/04/08"
output: 
  BiocStyle::html_document:
    toc: true
    toc_float: true
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, out.width = "100%", fig.align='center', 
                      message=FALSE, warning = FALSE)
options(width = 1200)
```

# Regress out gene signature
As we have now the gene signature for each cell type, we seek to correct for the observed bias to meet two objectives:

1. Remove technical artifacts. We will visualize this as an increased 'mixedness' of "affected" and "unaffected" clusters.
2. Preserve biological variability. Our purpose is that the only HVG that get lost are the ones present in the meta-signature.

To achieve that, we will regress out the time-score variable we computed for each cell in previous notebooks only for the genes in the meta-signatures. We will plot the tSNEs (with the HVG) before and after the correction to assess the overall performance.

This procedure is inspired by the correction of the cell cycle variability carried out in previous studies. We refer the reader to the following articles and Seurat vignette for a full understanding:

- http://www.sciencemag.org/cgi/pmidlookup?view=long&pmid=27124452
- https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6028641/#MOESM4
- https://satijalab.org/seurat/cell_cycle_vignette.html

```{r}
library(SingleCellExperiment)
library(scater)
library(scran)
library(Seurat)
library(EnsDb.Hsapiens.v86)
library(ggpubr)
library(org.Hs.eg.db)
library(biomaRt)
library(SC3)
library(pheatmap)
library(BiocStyle)
library(purrr)
library(ggrepel)
library(plotROC)
library(caret)
library(ROCR)
library(pheatmap)
library(org.Hs.eg.db)
library(GOstats)
library(VennDiagram)
library(ggplotify)
library(scales)
library(ggridges)
library(tidyverse)

source("bin/utils.R")

# Load SingleCellExperiment object
sce_10x <- readRDS("results/R_objects/sce_10x_scored.rds")

# Load list with metasignatures
metasignatures <- readRDS("results/R_objects/metasignatures.rds")

# Plot PCA all cells colored by label ("affected" or "unaffected")
cell_types <- c("CD4 T", "CD8 T", "NK", "Monocyte", "B")
seurat_10x <- Convert(from = sce_10x, to = "seurat")
seurat_10x <- SetAllIdent(seurat_10x, id = "cell_type")
seurat_list <- list(
  "CD4 T" = list(), 
  "CD8 T" = list(), 
  "NK" = list(), 
  "Monocyte" = list(), 
  "B" = list()
)
tsne_list <- seurat_list

for (cell_type in cell_types) {
  seurat_sub <- SubsetData(seurat_10x, ident.use = cell_type)
  for (i in c("not regressed", "regressed")) {
    seurat_curr <- ScaleData(seurat_sub, display.progress = FALSE)
    if (i == "regressed") {
      scaled_data_all <- seurat_curr@scale.data
      seurat_curr <- ScaleData(
        seurat_curr,
        genes.use = metasignatures[[cell_type]],
        vars.to.regress = "time_score",
        display.progress = FALSE
      )
      scaled_data_all[rownames(seurat_curr@scale.data), ] <- seurat_curr@scale.data
      seurat_curr@scale.data <- scaled_data_all
    } 
    seurat_curr <- FindVariableGenes(
      seurat_curr, 
      do.plot = FALSE, 
      display.progress = FALSE, 
      binning.method = "equal_frequency"
    )
    seurat_curr <- SetAllIdent(seurat_curr, id = "label")
    seurat_curr <- RunPCA(
      seurat_curr, 
      pc.genes = seurat_curr@var.genes
    )
    seurat_curr <- RunTSNE(seurat_curr)
    tsne_list[[cell_type]][[i]] <- TSNEPlot(seurat_curr)
    seurat_list[[cell_type]][[i]] <- seurat_curr
  }
}

df <- map(c("not regressed", "regressed"), function(regr) {
  tsne_list %>% 
    map(regr) %>% 
    map("data") %>% 
    bind_rows(.id = "cell_type")
})
names(df) <- c("original", "regressed")
df <- bind_rows(df, .id = "is_regressed")

tsne_regressed <- df %>% 
  mutate(cell_type = factor(cell_type, cell_types)) %>% 
  ggplot(aes_string("tSNE_1", "tSNE_2", color = "ident")) +
  geom_point(size = 0.6) +
  facet_grid(is_regressed ~ cell_type) +
  labs(x = "tSNE1", y = "tSNE2") +
  scale_color_manual("", values = c("red2", "royalblue1")) +
  theme_classic() +
  guides(colour = guide_legend(override.aes = list(size = 4)))
tsne_regressed
ggsave(
  filename = str_c("results/plots/", Sys.Date(), "_tsne_regressed.pdf"),
  plot = tsne_regressed,
  device = "pdf",
  width = 12,
  height = 7
)
saveRDS(tsne_regressed, file = "results/R_objects/tsne_regressed_gg.rds")

# Plot only CD4
tsne_regressed_cd4 <- df %>% 
  dplyr::filter(cell_type == "CD4 T") %>% 
  ggplot(aes(tSNE_1, tSNE_2, color = ident)) +
    geom_point(size = 0.25) +
    facet_grid(is_regressed ~ .) +
    theme_classic2() +
    labs(x = "tSNE1", y = "tSNE2") +
    scale_color_manual("", values = c("red2", "royalblue1")) 
saveRDS(tsne_regressed_cd4, file = "results/R_objects/tsne_regressed_cd4.rds")
```

# Supplementary figure
```{r eval=FALSE}
plot_list <- readRDS("results/R_objects/validation_plots.rds")
plot_list[[4]] <- tsne_regressed
tsne_regressed +  
    theme(legend.position = "none", 
          strip.background = element_rect(fill = NA))
plot_list <- map(plot_list, function(gg) {
  gg + 
    theme(legend.position = "none", 
          strip.background = element_rect(fill = "white"))
})
metasign_valid_gg <- ggarrange(
  plotlist = plot_list[1:3], 
  ncol = 1, 
  nrow = 3, 
  labels = "auto", 
  align = "v"
)
metasign_valid_gg2 <- ggarrange(
  plotlist = list(metasign_valid_gg, plot_list[[4]]), 
  ncol = 1, 
  nrow = 2, 
  heights = c(0.75, 0.25),
  labels = c("", "d")
)
ggsave(
  filename = str_c("doc/figures/R/", date, "_validation_figure.pdf"), 
  plot = metasign_valid_gg2, 
  width = 19, 
  height = 27,
  units = "cm"
)
metasign_valid_gg
```

# Session Info

```{r}
sessionInfo()
```
